✏️ Rename: news slider to news list [#73]
